Conflicts:
	src/app.js

+++ src/public/view/BudgetView.js
@@ -129,7 +129,7 @@ var BudgetView = Backbone.View.extend({
 
         this.$el.on('focusout', '#amount', function(e) {
             if(/^\d+(\.\d{1,2})?$/.test($('#amount').val())){
-                this.data.amount = parseInt($('#amount').val());
+                this.data.amount = parseFloat($('#amount').val());
                 self.isValid.amount = true;
                 self.$('#amountwarning').hide();
                 self.$('#amountrequest').hide();
